# [SW-636] Improve multisig performance