Table of dry runs called every 7 seconds

**with additional info** means that some pages demand additional info (`MultisigDetails`, `MultisigProposalDetails`), but we will have one more dry-run and only for active account
|Multisig info|Before|After (with additional info)|After (without additional info)|
|---|---|---|---|
|1 visible vault|**3** dry-runs | **2** dry-runs|**1** dry-run|
|10 visible vaults|**30** dry-runs | **11** dry-runs|**10** dry-runs|